refactor(swap): extract constants, split quote hooks, remove dead code
- Add swapConfig.ts with named constants (DEBOUNCE_MS, QUOTE_REFRESH_INTERVAL_MS,
  DEFAULT_SLIPPAGE_BPS, MAX_HOPS, SUSPICIOUS_VALUE_THRESHOLD_PCT, ORDER_HISTORY_LIMIT)
- Extract EVM_TOKEN_ADDRESS_TO_SYMBOL, EVM_TOKEN_ICON_MAP, STELLAR_TOKEN_ICON_MAP,
  and STELLAR_FALLBACK_CONTRACTS into lib/constants/tokenIcons.ts
- Split useSwapQuote (383 lines) into useEvmQuote + useStellarQuote + coordinator
- Move AbortController management into useStellarQuote; expose cancelStellarQuote
- Add sanitizeAmountInput and getTokenId helpers to swapUtils; simplify getTokenIcon
- useTokenSelection: replace inline getTokenId with swapUtils import, replace require()
  with static import for getTokenIcon, remove console.log statements
- useSwapState: remove dead setTokenInWrapper/setTokenOutWrapper with console.logs
- useSwapPrices: replace magic 10 with SUSPICIOUS_VALUE_THRESHOLD_PCT
- TokenInput/LimitOrderForm: use sanitizeAmountInput instead of inline regex
- OrderManagement: remove "coming soon" placeholder, fix {} as any with useWalletClient,
  use ORDER_HISTORY_LIMIT constant
- price.service.ts: remove ETH hardcoded fallback price
