+++ apps/web-app/src/features/swap/hooks/useSwapQuote.ts
@@ -1,23 +1,13 @@
-import { useState, useEffect, useRef, useCallback, useMemo } from "react";
-import { usePublicClient, useWalletClient } from "wagmi";
-import { parseUnits, formatUnits } from "viem";
+import { useState, useEffect, useRef, useCallback } from "react";
 import type { Token } from "@/lib/helpers/stellar/soroswap";
-import { isEVMToken, type EVMToken } from "@/lib/types/evmToken";
+import { hasApiKey } from "@/lib/helpers/stellar/soroswap";
+import { type EVMToken } from "@/lib/types/evmToken";
+import { useEvmQuote } from "./useEvmQuote";
+import { useStellarQuote } from "./useStellarQuote";
 import {
-  hasApiKey,
-  getQuote,
-  type QuoteRequest,
-} from "@/lib/helpers/stellar/soroswap";
-import {
-  getCowSwapQuote,
-  type CowSwapQuoteRequest,
-} from "@/lib/helpers/evm/cowswap";
-import { getTokensForChain } from "@/lib/constants/evmConfig";
-import {
-  formatSwapAmount,
-  fromSmallestUnit,
-} from "@/lib/helpers/stellar/swapUtils";
-import { QUOTE_TIMEOUT_MS } from "@/lib/constants/cowswapConfig";
+  DEBOUNCE_MS,
+  QUOTE_REFRESH_INTERVAL_MS,
+} from "@/features/swap/constants/swapConfig";
 
 export interface SwapQuoteState {
   amountOut: string;
@@ -40,19 +30,23 @@ export function useSwapQuote(
   const [amountOut, setAmountOut] = useState<string>("0.0");
   const [isLoadingQuote, setIsLoadingQuote] = useState<boolean>(false);
   const [apiKeyConfigured, setApiKeyConfigured] = useState<boolean>(false);
-  const publicClient = usePublicClient();
-  const { data: walletClient } = useWalletClient();
 
   const quoteTimeoutRef = useRef<NodeJS.Timeout | null>(null);
-  const abortControllerRef = useRef<AbortController | null>(null);
   const quoteIntervalRef = useRef<NodeJS.Timeout | null>(null);
 
-  const EVM_TOKENS = useMemo(
-    () => getTokensForChain(selectedEvmChainId),
-    [selectedEvmChainId]
+  const { fetchEvmQuote } = useEvmQuote(
+    amountIn,
+    tokenIn,
+    tokenOut,
+    selectedEvmChainId
+  );
+  const { fetchStellarQuote, cancelStellarQuote } = useStellarQuote(
+    amountIn,
+    tokenIn,
+    tokenOut,
+    apiKeyConfigured
   );
 
-  // Check if API key is configured on mount
   useEffect(() => {
     setApiKeyConfigured(hasApiKey());
   }, []);
@@ -74,208 +68,25 @@ export function useSwapQuote(
       return;
     }
 
-    // Define token symbols for both EVM and Stellar modes
-    const tokenInSymbol: string =
-      typeof tokenIn === "string"
-        ? tokenIn
-        : isEVMToken(tokenIn)
-          ? tokenIn.symbol || "ETH"
-          : "ETH";
-    const tokenOutSymbol: string =
-      typeof tokenOut === "string"
-        ? tokenOut
-        : isEVMToken(tokenOut)
-          ? tokenOut.symbol || "USDC"
-          : "USDC";
-
-    // EVM swap quote (CoW Swap)
-    if (swapMode === "evm" && publicClient && selectedEvmChainId) {
-      try {
-        const tokenInObj = EVM_TOKENS[tokenInSymbol];
-        if (!tokenInObj) {
-          setAmountOut("0.0");
-          setIsLoadingQuote(false);
-          return;
-        }
-
-        const amountInWei = parseUnits(
-          trimmedAmount,
-          tokenInObj.decimals
-        ).toString();
-
-        const quoteRequest: CowSwapQuoteRequest = {
-          tokenIn: tokenInSymbol,
-          tokenOut: tokenOutSymbol,
-          amountIn: amountInWei,
-        };
-
-        const quote = await getCowSwapQuote(
-          quoteRequest,
-          selectedEvmChainId,
-          publicClient,
-          walletClient
-        );
-
-        const tokenOutObj = EVM_TOKENS[tokenOutSymbol];
-        if (tokenOutObj) {
-          const amountOutFormatted = formatUnits(
-            BigInt(quote.amountOut),
-            tokenOutObj.decimals
-          );
-          setAmountOut(amountOutFormatted || "0.0");
-        } else {
-          // Fallback: format the amount even without token info
-          const amountOutStr = quote.amountOut.toString();
-          const amountOutFormatted = fromSmallestUnit(amountOutStr, 6);
-          setAmountOut(formatSwapAmount(amountOutFormatted, 6));
-        }
-      } catch (error) {
-        const errorMessage =
-          error instanceof Error ? error.message : String(error);
-        console.error("Quote error:", errorMessage);
-        setAmountOut("0.0");
-      } finally {
-        setIsLoadingQuote(false);
-      }
-      return;
-    }
-
-    // Stellar swap quote
-    if (swapMode === "stellar" && !apiKeyConfigured) {
-      setAmountOut("0.0");
-      setIsLoadingQuote(false);
-      return;
-    }
-
-    // Cancel previous request if exists
-    if (abortControllerRef.current) {
-      abortControllerRef.current.abort();
-    }
-
-    // Create new abort controller for this request
-    const abortController = new AbortController();
-    abortControllerRef.current = abortController;
-
-    // Store current input values to verify later
-    const currentAmountIn = trimmedAmount;
-    const currentTokenIn = tokenIn;
-    const currentTokenOut = tokenOut;
-
-    // For testing, ensure we have a reasonable amount
-    if (!trimmedAmount || parseFloat(trimmedAmount) <= 0) {
-      setAmountOut("0.0");
-      setIsLoadingQuote(false);
-      return;
-    }
-
-    setIsLoadingQuote(true);
     try {
-      const quoteRequest: QuoteRequest = {
-        assetIn: tokenIn as Token | string,
-        assetOut: tokenOut as Token | string,
-        amount: trimmedAmount,
-        tradeType: "EXACT_IN",
-        // Start with single protocol for testing
-        protocols: ["soroswap"],
-        slippageBps: 500, // 5% slippage (more generous)
-        maxHops: 1, // Reduce to direct routes only for testing
-      };
-
-      const quote = await getQuote(quoteRequest);
-
-      // Check if request was aborted or values changed
-      if (abortController.signal.aborted) {
-        return;
-      }
-
-      // Verify we're still looking at the same values
-      const currentAmountTrimmed = amountIn?.trim() || "";
-      const valuesMatch =
-        currentAmountIn === currentAmountTrimmed &&
-        currentTokenIn === tokenIn &&
-        currentTokenOut === tokenOut;
-
-      if (!valuesMatch) {
-        return;
-      }
-
-      // Process the quote result
-      if (quote && quote.amountOut) {
-        try {
-          const amountOutStr = quote.amountOut.toString();
-          const amountOutBigInt = BigInt(amountOutStr);
-
-          if (amountOutBigInt > BigInt(0)) {
-            const tokenOutObj = EVM_TOKENS[tokenOutSymbol];
-            let amountOutFormatted: string;
-
-            if (tokenOutObj) {
-              // EVM tokens - use token's decimals
-              amountOutFormatted = formatUnits(
-                amountOutBigInt,
-                tokenOutObj.decimals
-              );
-            } else {
-              // Stellar tokens use 7 decimals
-              amountOutFormatted = fromSmallestUnit(amountOutStr, 7);
-            }
-
-            const formatted = formatSwapAmount(amountOutFormatted, 6);
-            setAmountOut(formatted);
-          } else {
-            setAmountOut("0.0");
-          }
-        } catch {
-          setAmountOut("0.0");
-        }
-      } else {
-        setAmountOut("0.0");
-      }
-    } catch (error) {
-      // Ignore aborted requests
-      if (error instanceof Error && error.name === "AbortError") {
-        return;
-      }
-
-      const errorMessage =
-        error instanceof Error ? error.message : String(error);
-      const errorWithCode = error as { code?: string } | null | undefined;
-      const isTimeout =
-        errorMessage.includes("timeout") ||
-        errorMessage.includes("ECONNABORTED") ||
-        errorWithCode?.code === "ECONNABORTED";
-
-      if (isTimeout) {
-        setIsLoadingQuote(false);
-        return;
-      }
-
+      const result =
+        swapMode === "evm" ? await fetchEvmQuote() : await fetchStellarQuote();
+      setAmountOut(result ?? "0.0");
+    } catch {
       setAmountOut("0.0");
     } finally {
-      if (!abortController.signal.aborted) {
-        setIsLoadingQuote(false);
-      }
+      setIsLoadingQuote(false);
     }
-  }, [
-    address,
-    amountIn,
-    tokenIn,
-    tokenOut,
-    apiKeyConfigured,
-    swapMode,
-    publicClient,
-    selectedEvmChainId,
-    EVM_TOKENS,
-    walletClient,
-  ]);
+  }, [address, amountIn, swapMode, fetchEvmQuote, fetchStellarQuote]);
 
-  // Debounced live quote update
+  // Debounced quote trigger
   useEffect(() => {
     if (quoteTimeoutRef.current) {
       clearTimeout(quoteTimeoutRef.current);
       quoteTimeoutRef.current = null;
     }
 
+    cancelStellarQuote();
     setIsLoadingQuote(false);
 
     const trimmedAmount = amountIn?.trim() || "";
@@ -292,38 +103,27 @@ export function useSwapQuote(
 
     if (!isValid || !address) {
       setAmountOut("0.0");
-      setIsLoadingQuote(false);
       return;
     }
 
     if (swapMode === "stellar" && !apiKeyConfigured) {
       setAmountOut("0.0");
-      setIsLoadingQuote(false);
       return;
     }
 
     setAmountOut("0.0");
-
-    if (abortControllerRef.current) {
-      abortControllerRef.current.abort();
-      abortControllerRef.current = null;
-    }
-
     setIsLoadingQuote(true);
 
     quoteTimeoutRef.current = setTimeout(() => {
       void fetchLiveQuote();
-    }, 50);
+    }, DEBOUNCE_MS);
 
     return () => {
       if (quoteTimeoutRef.current) {
         clearTimeout(quoteTimeoutRef.current);
         quoteTimeoutRef.current = null;
       }
-      if (abortControllerRef.current) {
-        abortControllerRef.current.abort();
-        abortControllerRef.current = null;
-      }
+      cancelStellarQuote();
       setIsLoadingQuote(false);
     };
   }, [
@@ -334,9 +134,10 @@ export function useSwapQuote(
     apiKeyConfigured,
     fetchLiveQuote,
     swapMode,
+    cancelStellarQuote,
   ]);
 
-  // Auto-update quotes every 5 seconds for price fluctuations
+  // Auto-refresh quotes every interval while valid inputs exist
   useEffect(() => {
     if (
       address &&
@@ -350,7 +151,7 @@ export function useSwapQuote(
       }
       quoteIntervalRef.current = setInterval(() => {
         void fetchLiveQuote();
-      }, 5000);
+      }, QUOTE_REFRESH_INTERVAL_MS);
     } else {
       if (quoteIntervalRef.current) {
         clearInterval(quoteIntervalRef.current);
